Move debugger to an extension, simplifying non-debug REPL
This is a first step in reworking the debugger, which will have an
emphasis on moving more of the UI issues into usermode.

For starters it just moves the debugger into an extension, and removes
all of the debugger-related code from the C dispatch code behind the
console.  This means there's no conditional behavior in %host-main.c
based on things like "at_breakpoint" etc, and the only code that knows
about things like INTEGER! => FRAME! mappings is in the debug extension
itself.  This will hopefully facilitate things like BACKTRACE being
written entirely in usermode.

At the moment, the same HOST-CONSOLE routine is called by both the
main REPL loop and nested loops at breakpoints.  A simplistic hook
called INIT-DEBUGGER is used to tell the debugger extension about the
REPL function.

In order to further reduce the amount of C code needed behind the
REPL, the host no longer tolerates a return of an ERROR! from the REPL
function.  It will only accept a BLOCK! of code to run, and any error
printing the REPL wishes to do based on a failed LOAD or otherwise
must be done inside its own body.  It must then either reissue a new
prompt or just return some no-op code to have the same effect (the
latter is what is done here).

Also, the REPL function now gets an additional piece of information to
tell it if it's on its first cycle, or if a HALT occurred.  This means
that the last bit of UI which was still in the core for outputting
the string `[escape]` is now put into usermode.  This helped to tidy
%host-main.c even further, cleaning up a longstanding problem where
the output call could encounter a Ctrl-C while not actually running
any Rebol code.
