+++ src/core/c-port.c
@@ -231,9 +231,17 @@ REBOOL Wait_Ports(REBARR *ports, REBCNT timeout, REBOOL only)
         if (GET_SIGNAL(SIG_INTERRUPT)) {
             CLR_SIGNAL(SIG_INTERRUPT);
 
+            if (PG_Breakpoint_Hook == NULL)
+                fail (Error_Host_No_Breakpoint_Raw());
+
+            const REBOOL interrupted = TRUE;
+            const REBVAL*default_value = VOID_CELL;
+            const REBOOL do_default = FALSE;
+
             DECLARE_LOCAL (result);
-            if (Do_Breakpoint_Throws(result, TRUE, VOID_CELL, FALSE)) {
-                //
+            if ((*PG_Breakpoint_Hook)(
+                result, interrupted, default_value, do_default
+            )){
                 // !!! Consider Wait_Ports() callsites being re-engineered
                 // to be able to gracefully accept a throw generated by
                 // a RESUME from a breakpoint, e.g. `resume/do [throw 10]`.
