+++ src/extensions/debugger/ext-debugger-init.reb
@@ -1,0 +1,8 @@
+REBOL [
+    Title: "Debugger Extension"
+    name: 'Debugger
+    type: 'Extension
+    version: 1.0.0
+    license: {Apache 2.0}
+]
+
