+++ src/tools/file-base.r
@@ -40,7 +40,6 @@ core: [
     c-word.c
 
     ; (D)ebug
-    d-break.c
     d-crash.c
     d-dump.c
     d-eval.c
